Add readme for tests add a readme that could help when writing test for a new entity
