+++ test/orders/fixtures/__init__.py
@@ -1,0 +1,1 @@
+from .generate_order import *
